u8 -> u64

+++ src/types/events/voice_gateway/speaking.rs
@@ -33,7 +33,7 @@ bitflags! {
     ///
     /// See <https://discord.com/developers/docs/topics/voice-connections#speaking>
     #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, chorus_macros::SerdeBitFlags)]
-    pub struct SpeakingBitflags: u8 {
+    pub struct SpeakingBitflags: u64 {
         /// Whether we'll be transmitting normal voice audio
         const MICROPHONE = 1 << 0;
         /// Whether we'll be transmitting context audio for video, no speaking indicator
